requirements.txt
ipandas==1.4.0
numpy==1.20.0
streamlit==1.4.0
regex==2022.1.18
nltk==3.6.7
pickleshare==0.7.5
mglearn==0.1.9
scikit-learn==1.0.2